// src/app/model/link.ts
export interface Link {

   
    alias: string;

    link: string;

    shortLink: string;

    count?: number;
   
   
  //  constructor(alias:string,link:string,shortLink:string,count?:number){
  //   this.alias = alias;
  //   this.link = link;
  //   this.shortLink = shortLink;
  //   this.count = count != null || count != undefined ? count : 0;
  //  }


	

}